DISABLE: writing CachePath setting (instead the FASTBUILD_CACHE_PATH can be used). A cmake variable could be used to control this.
CLEANUP: custom command computation of outputs unified in new DetectCustomCommandOutputs method.
Improved (and simplified) logic to decide which command is common and which is config dependant, as well as which is a duplicate from an earlier command.
Duplicate commands no longer rely on aliases, instead the name of the earlier target is used directly.

Small tidying of the generated file, removing target name and output properties for global and utility targets that don't have any output.