FASTBuild: add support for CMAKE_VS_INCLUDE_INSTALL_TO_DEFAULT_BUILD / CMAKE_VS_INCLUDE_PACKAGE_TO_DEFAULT_BUILD variables
The ALL_BUILD target has to still be the same, as the INSTALL and PACKAGE rules depends on it.
So in case one of the above variables is set, a new final target is generated, ALL_INSTALL or ALL_PACKAGE or
ALL_INSTALL_PACKAGE depending on the selected options.

Also add a '-FASTBuild' suffix to the generated solution names to make it easier to distinguish "normal" solutions
from ones where FASTBuild is used behind the scenes. This behavior probably should be configurable with a variable.