Fastbuild: add a cmFastbuildLinkLineComputer class, overriding ComputeLinkLibs
Added virtual on cmLinkLineComputer::ComputeLinkLibs.

This allows us to put some of the libraries to link in the .Libraries
FASTBuild var, instead of passing everything in .LinkerOptions
(it was already like this in the old version of the Fastbuild generator
but the recent upgrade to CMake v3.11.4 modified the behavior.
After the upgrade all libraries were put in the .LinkerOptions,
but I had compilation issues with symbols defined multiple times
on some projects)